reformatting install_packages
